Add files via upload
So a lot of things changed, I realized what was causing the discrepancy in the graphs and eigenvalues and has been corrected. The problem was focused around the angular frequency not being coded correctly and has been fixed. There was also a factor of k=m_0=1 that was messing the entire code up and has been fixed. This was the reason I had a scaling factor and now all scaling factors have been reduced or changed to 1 as they should be. Hope this comment explains some stuff.