Add files via upload
Got to work on this one right after assignment 3, thought this was a bit easier was able to get through it relatively quickly. Did some work with PDFs in modern physics so it helped a lot. Note the obnoxiously long print statements, I wanted to save lines and I have not tried fitting everything in one print statement, I apologize. But surprisingly is very satisfying.